wip: style js refactor
